Libraries, agent, extension and integrations
Update the terminology for the Ruby gem and Elixir package to be
referred to as libraries. The agent is the system's agent. The extension
is the C extension used by the Ruby gem and Elixir package. Integrations
refer to either integrations the Ruby/Elixir library have or a
third-party integration such as Slack.