Add ca-certificates as dep for Debian systems
`curl` will also install ca-certificates for the Elixir package, but
it's good to mention it separately if we ever get around to removing the
curl dependency.

The package is not necessarily needed for the Ruby gem either as we
bundle our own certificates, but it can't hurt to add it as a fallback.

Closes #141