Fix the link to the config page on ignore actions
For some reason the Ruby page linked to the Elixir page. The link is
relative, but middleman generates the link to the Elixir page.

Use a relative path that navigates the directory structure to
(hopefully) force the correct page to be linked.

Reported in https://github.com/appsignal/appsignal-server/issues/3470