Update middleman to the latest 3.x release
Why this upgrade? Because every time I change something in `config.rb`
or files `config.rb` requires the server (`middleman server`) gets stuck
and it has to be stopped and started fresh. If the server is left alone
the page either does not update with new content or returns an error
that the server can't be reached.

Update middleman gems and dependencies. Updated middleman to the latest
3.x series before upgrading to 4.x, because it's an easier upgrade path
preparing us for an upgrade to 4.x.

Changes to syntax highlighting
==============================

During the upgrade the syntax highlighting broke slightly. To fix it the
following changes were made.

1. The generate code block HTML is set up a little differently. It wraps
  code block with a `pre` element higher in the element. It also adds a
  `.highlight` class to this `pre`, which creates some issues with the
  new code blocks adding more spacing.
  To fix this configure the CSS class of the `pre` element to an unused
  `code-block` class in `config.rb`.

2. New pygments, code highlighting, CSS is loaded in based on the
  following command:

`rougify style base16.solarized.light > source/stylesheets/vendor/pygments.css.scss`

This stylesheet is not the exact same as used before, since it's based
on Base16, so I'll tweak the code highlighting to match the old style in
a future commit.

Alternative we could use the following helper to generate them from
rouge, but that would mean we can no longer directly customize the CSS.

```
// /source/stylesheets/vendor/pygments.css.erb
<%= Rouge::Themes::Base16::Solarized.render(:scope => '.highlight') %>
```