Update default value for request_headers option
This was changed after the `.0` releases of the Ruby gem and Elixir
package to an actual list of values rather than adding the config to the
config file for the app. When a user wants to change this value they
need to copy the default value and add their own values.