Update data life cycle page
Update to the new integrations structure using the agent. I think it was
still taking about how Ruby gem version 0.x worked.

Rename the request lifecycle page to be about "data" as a more generic
term to also cover background jobs, custom transactions and custom/host
metrics.

This change also describe how Anomaly detection processing is done as
this question comes up in support every once in a while. This is a
simplified version leaving out a lot of details to keep the process
understandable.

An entry detailing how to set up NTP, or at least a separate section
detailing the problem which we can refer to from other pages as well,
will follow later.